fix: move concat & concat message extra by ConcatItems
